split interfaces
