+++ ng2-components/ng2-activiti-form/src/components/widgets/dynamic-table/cell-validator.model.ts
@@ -15,9 +15,15 @@
  * limitations under the License.
  */
 
-export { DataTableAdapter } from './datatable-adapter';
-export { ObjectDataColumn, ObjectDataRow, ObjectDataTableAdapter } from './object-datatable-adapter';
-export { DataRow } from './data-row.model'
-export { DataRowEvent } from './data-row-event.model';
-export { DataColumn } from './data-column.model';
-export { DataSorting } from './data-sorting.model';
+/* tslint:disable:component-selector  */
+
+import { DynamicTableColumn } from './dynamic-table-column.model'
+import { DynamicTableRow } from './dynamic-table-row.model'
+import { DynamicRowValidationSummary } from './dynamic-row-validation-summary.model'
+
+export interface CellValidator {
+
+    isSupported(column: DynamicTableColumn): boolean;
+    validate(row: DynamicTableRow, column: DynamicTableColumn, summary?: DynamicRowValidationSummary): boolean;
+
+}
