+++ ng2-components/ng2-activiti-form/src/components/widgets/dynamic-table/dynamic-table-row.model.ts
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-export { DataTableAdapter } from './datatable-adapter';
-export { ObjectDataColumn, ObjectDataRow, ObjectDataTableAdapter } from './object-datatable-adapter';
-export { DataRow } from './data-row.model'
-export { DataRowEvent } from './data-row-event.model';
-export { DataColumn } from './data-column.model';
-export { DataSorting } from './data-sorting.model';
+/* tslint:disable:component-selector  */
+
+export interface DynamicTableRow {
+    isNew: boolean;
+    selected: boolean;
+    value: any;
+}
