+++ ng2-components/ng2-activiti-form/src/components/widgets/dynamic-table/editors/dropdown/dropdown.editor.ts
@@ -20,7 +20,10 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { LogService } from 'ng2-alfresco-core';
 import { FormService } from './../../../../../services/form.service';
-import { DynamicTableColumn, DynamicTableColumnOption, DynamicTableModel, DynamicTableRow } from './../../dynamic-table.widget.model';
+import { DynamicTableModel } from './../../dynamic-table.widget.model';
+import { DynamicTableRow } from './../../dynamic-table-row.model';
+import { DynamicTableColumn } from './../../dynamic-table-column.model';
+import { DynamicTableColumnOption } from './../../dynamic-table-column-option.model';
 
 @Component({
     selector: 'adf-dropdown-editor',
