+++ ng2-components/ng2-activiti-form/src/components/widgets/dynamic-table/editors/row.editor.ts
@@ -18,7 +18,10 @@
  /* tslint:disable:component-selector  */
 
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { DynamicRowValidationSummary, DynamicTableColumn, DynamicTableModel, DynamicTableRow } from './../dynamic-table.widget.model';
+import { DynamicTableModel } from './../dynamic-table.widget.model';
+import { DynamicTableRow } from './../dynamic-table-row.model';
+import { DynamicTableColumn } from './../dynamic-table-column.model';
+import { DynamicRowValidationSummary } from './../dynamic-row-validation-summary.model';
 
 @Component({
     selector: 'row-editor',
