+++ ng2-components/ng2-activiti-form/src/events/validate-dynamic-table-row.event.ts
@@ -15,7 +15,9 @@
  * limitations under the License.
  */
 
-import { DynamicRowValidationSummary, DynamicTableRow } from '../components/widgets/dynamic-table/dynamic-table.widget.model';
+import { DynamicTableRow } from '../components/widgets/dynamic-table/dynamic-table-row.model';
+import { DynamicRowValidationSummary } from '../components/widgets/dynamic-table/dynamic-row-validation-summary.model';
+
 import { FormFieldModel, FormModel } from './../components/widgets/core/index';
 import { FormFieldEvent } from './form-field.event';
