+++ ng2-components/ng2-alfresco-datatable/src/components/datatable/datatable.component.ts
@@ -22,7 +22,12 @@ import {
 import { MatCheckboxChange } from '@angular/material';
 import { DataColumnListComponent } from 'ng2-alfresco-core';
 import { Observable, Observer, Subscription } from 'rxjs/Rx';
-import { DataColumn, DataRow, DataRowEvent, DataSorting, DataTableAdapter } from '../../data/datatable-adapter';
+import { DataTableAdapter } from '../../data/datatable-adapter';
+import { DataRow } from '../../data/data-row.model';
+import { DataRowEvent } from '../../data/data-row-event.model';
+import { DataSorting } from '../../data/data-sorting.model';
+import { DataColumn } from '../../data/data-column.model';
+
 import { ObjectDataRow, ObjectDataTableAdapter } from '../../data/object-datatable-adapter';
 import { DataCellEvent } from './data-cell.event';
 import { DataRowActionEvent } from './data-row-action.event';
