+++ ng2-components/ng2-alfresco-datatable/src/data/data-column.model.ts
@@ -15,9 +15,16 @@
  * limitations under the License.
  */
 
-export { DataTableAdapter } from './datatable-adapter';
-export { ObjectDataColumn, ObjectDataRow, ObjectDataTableAdapter } from './object-datatable-adapter';
-export { DataRow } from './data-row.model'
-export { DataRowEvent } from './data-row-event.model';
-export { DataColumn } from './data-column.model';
-export { DataSorting } from './data-sorting.model';
+import { TemplateRef } from '@angular/core';
+
+export interface DataColumn {
+    key: string;
+    type: string; // text|image|date
+    format?: string;
+    sortable?: boolean;
+    title?: string;
+    srTitle?: string;
+    cssClass?: string;
+    template?: TemplateRef<any>;
+    formatTooltip?: Function;
+}
