+++ ng2-components/ng2-alfresco-datatable/src/data/data-row-event.model.ts
@@ -15,9 +15,18 @@
  * limitations under the License.
  */
 
-export { DataTableAdapter } from './datatable-adapter';
-export { ObjectDataColumn, ObjectDataRow, ObjectDataTableAdapter } from './object-datatable-adapter';
-export { DataRow } from './data-row.model'
-export { DataRowEvent } from './data-row-event.model';
-export { DataColumn } from './data-column.model';
-export { DataSorting } from './data-sorting.model';
+import { BaseUIEvent } from 'ng2-alfresco-core';
+import { DataRow } from './data-row.model'
+
+export class DataRowEvent extends BaseUIEvent<DataRow> {
+
+    sender: any;
+
+    constructor(value: DataRow, domEvent: Event, sender?: any) {
+        super();
+        this.value = value;
+        this.event = domEvent;
+        this.sender = sender;
+    }
+
+}
