+++ ng2-components/ng2-alfresco-datatable/src/data/object-datatable-adapter.ts
@@ -19,7 +19,12 @@ import { DatePipe } from '@angular/common';
 import { TemplateRef } from '@angular/core';
 
 import { ObjectUtils, TimeAgoPipe } from 'ng2-alfresco-core';
-import { DataColumn, DataRow, DataSorting, DataTableAdapter } from './datatable-adapter';
+import { DataTableAdapter } from './datatable-adapter';
+import { DataColumn } from './data-column.model';
+import { DataRow } from './data-row.model';
+import { DataSorting } from './data-sorting.model';
+
+
 
 // Simple implementation of the DataTableAdapter interface.
 export class ObjectDataTableAdapter implements DataTableAdapter {
