+++ ng2-components/ng2-alfresco-documentlist/src/data/image-resolver.model.ts
@@ -15,9 +15,7 @@
  * limitations under the License.
  */
 
-export { DataTableAdapter } from './datatable-adapter';
-export { ObjectDataColumn, ObjectDataRow, ObjectDataTableAdapter } from './object-datatable-adapter';
-export { DataRow } from './data-row.model'
-export { DataRowEvent } from './data-row-event.model';
-export { DataColumn } from './data-column.model';
-export { DataSorting } from './data-sorting.model';
+import { DataColumn, DataRow } from 'ng2-alfresco-datatable';
+
+
+export type ImageResolver = (row: DataRow, column: DataColumn) => string;
