+++ ng2-components/ng2-alfresco-documentlist/src/data/row-filter.model.ts
@@ -15,9 +15,6 @@
  * limitations under the License.
  */
 
-export { DataTableAdapter } from './datatable-adapter';
-export { ObjectDataColumn, ObjectDataRow, ObjectDataTableAdapter } from './object-datatable-adapter';
-export { DataRow } from './data-row.model'
-export { DataRowEvent } from './data-row-event.model';
-export { DataColumn } from './data-column.model';
-export { DataSorting } from './data-sorting.model';
+import { ShareDataRow } from './share-data-row.model';
+
+export type RowFilter = (value: ShareDataRow, index: number, array: ShareDataRow[]) => any;
